Automatically size to desktop width/height.  Fix an extremely stupid bug with FromZeroOne and ToZeroOne - source parameter was never getting read.  Why did this work on OSX?  I don't believe it just appeared to. Stack must have aliased old copy.

+++ Render/Shaders/light.h
@@ -1,7 +1,9 @@
+// -*- mode: glsl; tab-width: 4; c-basic-offset: 4; -*-
+
 
 struct PointLight
 {
-    float m_TypePad;
+    int m_TypePad;
     float m_Range;
     int m_PadB;
     int m_PadC;
